Move methods in IdentifierList to be more generic
Move node related methods to node, make registration inheret from node
and remove get_serializer method, instead let each class define its own
seializer.
